requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
altair
google-generativeai